Helpful warnings in development
Adds a new module `createHelper`. In production, it is an alias
for lodash.curry. In development, it is a custom curry function
that adds special flags to partially applied higher-order component
helpers. The `compose` module has been updated to use these flags
to detect if a helper has been applied with insufficient arguments.