Fix mapPropsStream applying transforms twice (#207)
Because mapPropsStream composes componentFromStream, the global
observable transforms were being applied twice. We'll fix by
exporting a factory version of each observable util (including
componentFromStream) that allows you opt-out of the global
transform.

This was not caught earlier because the tests for mapPropsStream
use RxJS 5, which is not affected by double transforms because
its observables are ES observables, anyway. To protect against
regressions, we'll switch to testing with RxJS4 for now.