Misc fixes & improvements (#267)
* Fix mapPropsStream's example

It seems like in the docs, `Observable` refers to `Rx.Observable`.

In Rx4 and Rx5, `Rx.Observable.interval` emits numbers, not objects.

* Advertise `withProps`

* Clean-up some comments

* Fix links to the React documentation

* Improve renderComponent's doc

Since faaafb0 (#227), branch's third argument defaults to the identity
function. This is a nice addition and we probably want to advertise it
in renderComponent's doc.

* Remove dead code

Since 390697a (#180, 2016-05-21), branch()'s `{Left,Right}Component`
internal instance variables are not used anymore.